pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "report_agent"
version = "0.0.1"
dependencies = [
  "langchain_experimental",
  "langchain_aws",
  "langchain_openai",
  "streamlit",
  "tabulate",
  "mcp",
  "mysql-connector-python"
]
requires-python = ">=3.9"
authors = [
  {name = "n3xtchen", email = "echenwen@gmail.com"},
]
maintainers = [
  {name = "n3xtchen", email = "echenwen@gmail.com"}
]
description = "Agent For Report"
readme = "README.md"
license = "MIT"
license-files = ["LICEN[CS]E.*"]
keywords = ["llm", "agent", "langchain"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]

[project.optional-dependencies]
dev = [
  "pip-tools",
  "pytest",
  "pytest_asyncio",
]

[project.urls]
Homepage = "https://n3xtchen.github.io/n3xtchen/"
Repository = "https://github.com/n3xtchen/report_agent"
"Bug Tracker" = "https://github.com/n3xtchen/report_agent/issues"

[project.scripts]
report_agent_server="script:run"

[tool.pytest.ini_options]
addopts = "-s"
pythonpath = [
  "src"
]
asyncio_default_fixture_loop_scope = "function"
